EW-949 adding registration with link from admin
