+++ cypress/e2e/login_management/registerStudent.feature
@@ -1,0 +1,45 @@
+@pr
+@release
+@stable_test
+Feature: Register an already existing student with the registration link
+
+    As a student, I want to register in the SVS
+
+    Scenario Outline: Setup
+        Given I am logged in as '<user>' on '<env>'
+        When Created student 'John' 'Student' with email 'john.student@mail.tld'
+        Then I am on the students management page
+
+    Examples:
+        | env | user       |
+        | dbc | admin1_dbc |
+        | brb | admin1_brb |
+        | nbc | admin1_nbc |
+
+    Scenario Outline: Student can registers in the SVS with the registration link
+        Given I am logged in as '<user>' on '<env>'
+        And I am on the students management page
+        Given Going to student edit page for 'John' with email 'john.student@mail.tld'
+        And Going to student registration page
+        And I am registering as student with email 'john.student@mail.tld' on '<env>'
+        Then I am able to logout
+
+    Examples:
+        | env | user       |
+        | dbc | admin1_dbc |
+        | brb | admin1_brb |
+        | nbc | admin1_nbc |
+
+    Scenario Outline: Teardown
+        Given I am logged in as '<user>' on '<env>'
+        And I am on the students management page
+        And Going to student edit page for 'John' with email 'john.student@mail.tld'
+        When Deleting the user
+        Then I am on the students management page
+
+    Examples:
+        | env | user       |
+        | dbc | admin1_dbc |
+        | brb | admin1_brb |
+        | nbc | admin1_nbc |
+
